upgrade to koin 3
Former-commit-id: a35d4cc099d0b20df8742200fc291cc5760e3de9 [formerly e8ef08f6a5b99b9a3abaab2739e3dd87d901feaa]
Former-commit-id: 3191cf4baa11835f9ecb375eb2e57476ce1300e7

+++ id/src/main/java/com/simprints/id/activities/settings/fragments/moduleselection/tools/ModuleChipHelper.kt
@@ -10,7 +10,7 @@ import com.google.android.material.shape.CornerFamily
 import com.google.android.material.shape.ShapeAppearanceModel
 import com.simprints.id.R
 import com.simprints.id.moduleselection.model.Module
-import org.jetbrains.anko.dimen
+import splitties.resources.dimen
 
 class ModuleChipHelper(private val context: Context, private val listener: ChipClickListener) {
 
@@ -44,7 +44,7 @@ class ModuleChipHelper(private val context: Context, private val listener: ChipC
     private fun createChipDrawable(): ChipDrawable {
         return ChipDrawable.createFromResource(context, R.xml.module_selection_chip).apply {
             setTextAppearanceResource(R.style.SimprintsStyle_TextView_Chip)
-            val cornerSize = context.dimen(R.dimen.chip_corner_size_module_selection).toFloat()
+            val cornerSize = context.dimen(R.dimen.chip_corner_size_module_selection)
             shapeAppearanceModel =
                 ShapeAppearanceModel.builder().setAllCorners(CornerFamily.CUT, cornerSize).build()
         }
